refactor: recipe update
